Enable users to initiate a shiny update from plugin detail modal
The plugin detail modal can contain a link to update a plugin.  When it does, we should initiate a shiny update.

This relies upon postMessage which isn't available in all browsers, specifically it isn't in IE versions below 8 so this is going to be a progressive enhancement that some small percentage of users will miss out on.  These are the same users that can't use the customizer.

Fixes #31739

Built from https://develop.svn.wordpress.org/trunk@32062