Customizer: Prevent loss of `walker` and `fallback_cb` args for `wp_nav_menu`.
These args only need to be cleared out when exported to JavaScript, when they are not JSON-serializable. So the filter now clears these when gathering args for exporting to JS, but otherwise now leaves the original values to be passed through to `wp_nav_menu()`.

Fixes #32781.

Built from https://develop.svn.wordpress.org/trunk@33131