MS: Populate `public` on empty `$current_blog` during subdomain activation.
Activation of a subdomain site is done through that new site's address. This address does not exist in the `wp_blogs` table until activation is complete.

In this case we need to make sure `public` is populated to avoid a PHP notice.

Props uglyrobot.
Fixes #24760.

Built from https://develop.svn.wordpress.org/trunk@35782